refactor(nms): Move types out of fbc_js_core
