+++ nms/app/components/EditUserDialog.js
@@ -31,7 +31,7 @@ import React, {useCallback, useContext, useMemo, useState} from 'react';
 import Select from '@material-ui/core/Select';
 import TextField from '@material-ui/core/TextField';
 import renderList from '../util/renderList';
-import {UserRoles} from '../../shared/types';
+import {UserRoles} from '../../shared/roles';
 import {makeStyles} from '@material-ui/styles';
 
 export type EditUser = {
