+++ nms/app/components/__tests__/NetworkSelector-test.js
@@ -20,7 +20,7 @@ import NetworkContext from '../context/NetworkContext';
 import NetworkSelector from '../NetworkSelector';
 import React from 'react';
 import {AppContextProvider} from '../../../app/components/context/AppContext';
-import {LTE} from '../../../fbc_js_core/types/network';
+import {LTE} from '../../../shared/types/network';
 import {MemoryRouter} from 'react-router-dom';
 import {SnackbarProvider} from 'notistack';
