+++ nms/app/components/admin/AddNetworkDialog.js
@@ -31,8 +31,8 @@ import TextField from '@material-ui/core/TextField';
 import axios from 'axios';
 
 import nullthrows from '../../../shared/util/nullthrows';
-import {AllNetworkTypes, XWFM} from '../../../fbc_js_core/types/network';
-import {CWF, FEG, FEG_LTE} from '../../../fbc_js_core/types/network';
+import {AllNetworkTypes, XWFM} from '../../../shared/types/network';
+import {CWF, FEG, FEG_LTE} from '../../../shared/types/network';
 import {makeStyles} from '@material-ui/styles';
 import {triggerAlertSync} from '../../state/SyncAlerts';
 import {useEnqueueSnackbar} from '../../../app/hooks/useSnackbar';
