+++ nms/app/components/admin/EditNetworkDialog.js
@@ -22,7 +22,7 @@ import LoadingFillerBackdrop from '../LoadingFillerBackdrop';
 import MagmaV1API from '../../../generated/WebClient';
 
 import useMagmaAPI from '../../../api/useMagmaAPI';
-import {CWF, FEG} from '../../../fbc_js_core/types/network';
+import {CWF, FEG} from '../../../shared/types/network';
 import {useEnqueueSnackbar} from '../../../app/hooks/useSnackbar';
 import {useParams} from 'react-router-dom';
