+++ nms/app/components/main/Index.js
@@ -16,11 +16,11 @@
 
 import MagmaV1API from '../../../generated/WebClient';
 
-import {FEG, coalesceNetworkType} from '../../../fbc_js_core/types/network';
+import {FEG, coalesceNetworkType} from '../../../shared/types/network';
 import {FEGContextProvider} from '../feg/FEGContext';
 import {LteContextProvider} from '../lte/LteContext';
 import {VersionContextProvider} from '../context/VersionContext';
-import type {NetworkType} from '../../../fbc_js_core/types/network';
+import type {NetworkType} from '../../../shared/types/network';
 import type {Theme} from '@material-ui/core';
 import type {
   network_id,
