+++ nms/app/views/organizations/OrganizationUserDialog.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Select from '@material-ui/core/Select';
 
 import {AltFormField, PasswordInput} from '../../../app/components/FormField';
-import {UserRoles} from '../../../shared/types';
+import {UserRoles} from '../../../shared/roles';
 import {makeStyles} from '@material-ui/styles';
 
 const useStyles = makeStyles(_ => ({
