+++ nms/app/views/organizations/OrganizationUsersTable.js
@@ -21,7 +21,7 @@ import withAlert from '../../components/Alert/withAlert';
 import type {EditUser} from './OrganizationEdit';
 import type {WithAlert} from '../../components/Alert/withAlert';
 
-import {UserRoles} from '../../../shared/types';
+import {UserRoles} from '../../../shared/roles';
 import {useEnqueueSnackbar} from '../../../app/hooks/useSnackbar';
 import {useParams} from 'react-router-dom';
 import {useState} from 'react';
