+++ nms/server/features.js
@@ -16,7 +16,7 @@
 
 import type {ExpressRequest, ExpressResponse, NextFunction} from 'express';
 import type {FBCNMSRequest} from './auth/access';
-import type {FeatureID} from '../fbc_js_core/types/features';
+import type {FeatureID} from '../shared/types/features';
 
 import {FeatureFlag} from '../fbc_js_core/sequelize_models';
