+++ nms/server/host/routes.js
@@ -16,7 +16,7 @@
 
 import type {ExpressResponse} from 'express';
 import type {FBCNMSRequest} from '../auth/access';
-import type {FeatureID} from '../../fbc_js_core/types/features';
+import type {FeatureID} from '../../shared/types/features';
 
 import Sequelize from 'sequelize';
 import asyncHandler from '../util/asyncHandler';
