+++ nms/shared/types/network.js
