+++ nms/shared/types/tabs.js
